+++ Zuenok/BookLibraryCRUD/BookLibraryCRUD/BookLibraryCRUD.csproj
@@ -5,4 +5,8 @@
     <TargetFramework>netcoreapp2.1</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="App_Data\" />
+  </ItemGroup>
+
 </Project>
