+++ Zuenok/BookLibraryCRUD/BookLibraryCRUD/Models/Book.cs
@@ -1,0 +1,10 @@
+﻿using System.Runtime.Serialization;
+
+namespace BookLibraryCRUD {
+    [DataContract]
+    public class Book
+    {
+        [DataMember] public int Id { get; set; }
+        [DataMember] public string Title { get; set; }
+    }
+}
