+++ Zuenok/BookLibraryCRUD/BookLibraryCRUD/Models/IRepository.cs
@@ -1,0 +1,7 @@
+﻿namespace BookLibraryCRUD {
+    public interface IRepository<out T>
+    {
+        T Get(int id);
+        T GetLast();
+    }
+}
