+++ Zuenok/BookLibraryCRUD/BookLibraryCRUD/Models/LibraryContext.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace BookLibraryCRUD {
+    public class LibraryContext
+    {
+        public IList<Book> Books { get; set; }
+    }
+}
